Updated Code for Cleaned 'InvertResults'
The code has been updated to provide the steps to clean the USGS BioData 'InvertResults' dataset. Given the different subsampling techniques used for the NAWQA sampling, this code will identify and reduce duplicates, "lab-large rare" entries, and then combine all duplicate taxon entries for a given site-by-lifestage (summed Abundance). 'TotalRows' is the resulting cleaned dataset that is used for analyses.